Reorganized code into services and systems folders; renamed objects to entities
