+++ main.lua
@@ -1,10 +1,11 @@
-require('keyboard')
 require('world')
-objects = require('objects')
-systems = require('systems')
-require('sounds')
-Camera = require('camera')
+require('entities')
+require('systems')
 require('state')
+
+require('/services/debug')
+require('/services/sounds')
+require('/services/keyboard')
 
 love.load = function()
     seconds = 0
@@ -13,60 +14,8 @@ love.load = function()
     love.window.setMode(800, 600)
     sounds.loadSounds()
 
-    systems.call('SpawnObjects')
-end
-
-function debug()
-    -- Displays certain values useful for debugging
-    function roundOff(value)
-        return math.floor(value * 100) / 100
-    end
-
-    -- View all objects currently in the entity component system
-    if love.keyboard.isScancodeDown('n') then
-        for index, data in ipairs(objects) do
-            print(index)
-
-            for key, value in pairs(data) do
-                print('\t', key, value)
-            end
-        end
-    elseif love.keyboard.isScancodeDown('m') then
-        print('=================================')
-    end
-
-    entity = systems.grab('player')
-    if entity.name == 'player' then
-        love.graphics.setColor({1, 1, 1, 1})
-        local clock_display = 'Time: ' .. roundOff(seconds)
-        love.graphics.print(clock_display, pos_x, pos_y, 0, 2, 2)
-        
-        pos_x = state.camera.pos_x
-        pos_y = state.camera.pos_y
-        local playerPosition = 'X/Y Pos: ' .. roundOff(entity.body:getX()) .. '/' .. roundOff(entity.body:getY())
-        love.graphics.print(playerPosition, pos_x, pos_y + 25, 0, 2, 2)
-        local currentVelocity = {entity.body:getLinearVelocity()}
-        local playerVelocity = 'X/Y Vel: ' .. roundOff(currentVelocity[1]).. '/' .. roundOff(currentVelocity[2])
-        love.graphics.print(playerVelocity, pos_x, pos_y + 50, 0, 2, 2)
-
-        local ratios = 'Gravity: ' .. forcex .. ':' .. forcey
-        love.graphics.print(ratios, pos_x, pos_y + 75, 0, 2, 2)
-
-        local campos = 'Camera: ' .. pos_x .. ':' .. pos_y
-        love.graphics.print(campos, pos_x, pos_y + 100, 0, 2, 2)
-
-        -- Direction line
-        local lastColor = {love.graphics.getColor()}
-        love.graphics.setColor({1, 0, 0, 1})
-        for _,value in pairs(objects) do
-            local currentVelocity = {value.body:getLinearVelocity()}
-            local velocityArrow = {value.body:getX() + currentVelocity[1], value.body:getY() + currentVelocity[2]}
-            love.graphics.line(value.body:getX(), value.body:getY(), velocityArrow[1], velocityArrow[2])
-        end
-        
-        velocityArrow = {entity.body:getX() + currentVelocity[1], entity.body:getY() + currentVelocity[2]}
-        love.graphics.setColor(lastColor)
-        love.graphics.print('VelocityArrow: ' .. roundOff(velocityArrow[1]) .. '/' .. roundOff(velocityArrow[2]), pos_x, pos_y + 125, 0, 2, 2)
+    for _, entity in ipairs(entities) do
+        systems.SpawnEntities(entity)
     end
 end
 
@@ -81,7 +30,9 @@ end
 love.draw = function()
     Camera.set()
 
-    systems.call('DrawObjects')
+    for _, entity in ipairs(entities) do
+        systems.DrawEntities(entity)
+    end
 
     if state.debugOn then debug() end
 
@@ -105,9 +56,9 @@ love.update = function(dt)
     if state.paused == false then
         world:update(dt)
         seconds = seconds + dt
-        for _, entity in ipairs(objects) do
+        for _, entity in ipairs(entities) do
             systems.ControlPlayer(entity)
-            systems.UpdateObjects(entity)
+            systems.UpdateEntities(entity)
             systems.Gravitate(entity, entity)
             systems.UpdateCamera(entity)
         end
