+++ systems/UpdateEntities.lua
@@ -1,0 +1,12 @@
+-- Call any update functions that need to be updated
+
+System = require('/lib/system')
+
+UpdateEntities = System(
+    {'update'},
+    function(updateFunction)
+        updateFunction()
+    end
+)
+
+return UpdateEntities
